core: Remove the Findbugs profiles
Currently we have two Maven profiles related to Findbugs, one named
"findbugs" that adds the project specific exclussions, and another one
named "findbugs-general" that also adds the shared exclussions. As both
exclussions need to be always enabled these profiles aren't needed,
instead the required settings can be specified in the Findbugs plugin,
inside the "plugins" section of the root POM.

After applying this patch the invocation of Findbugs will just be the
following:

  $ mvn findbugs:findbugs

The "-Pfindbugs-general" option is no longer required.

Change-Id: Ife1fafa2b6b7db4ee12d61835319059396a1f68e
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>