engine: Import clone VM without copy-collapse
The user should be able to import and clone a VM without snapshots
without setting copy-collapse to true via REST.

Also, when importing a VM without snapshots that is based on a template
the copy-collapse parameter should not be ignored.

The flow in this case in  CopyImageGroupCommand should be the same as
when copy-collapse is set to true.

Change-Id: I8b3f2096341bef7a795813c304aebff3a0ee279e
Bug-Url: https://bugzilla.redhat.com/1147011
Signed-off-by: Fred Rolland <frolland@redhat.com>