restapi: Add "StructMember" and "MemberMethod" classes
Currently the struct type attributes and links have many, in fact almost
all in common, except their name. In order to reduce the code
duplication that this generates this patch introduces a new
"StructMember" base class to contain all the common fields and methods.

The same happens with service methods and locators. This patch
introduces a new "ServiceMember" class to simplify that.

Change-Id: I56de044533340dc4751f5fb7f234e41c8071ad99
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>