core: live snapshot - create VM metadata in RAW
Currently, VM metadata file is created in COW
format during live snapshot flow. However, dumping
the data (using pickle on snapshot verb) transforms
the file to RAW.

As a result, we get a miscorrelation
between the .meta file (which still has COW as a format),
while the file itself is RAW (verified using 'qemu info').
Hence, the VM metadata file should be created as RAW format
in the first place instead of COW.

Change-Id: I12c1d657661c86f129df92e633f0438d778d6147
Bug-Url: https://bugzilla.redhat.com/1284580
Signed-off-by: Daniel Erez <derez@redhat.com>