core: do not pass the context automatically on rollback
In change I310f5f77fff78b3232ee77fe63791425fd521516
CommandBase.attemptRollback() was changed so that instead of accepting a
callback as a parameter it automatically clones the current command
context and passes it.
That changed the previous behavior, as the using commands didn't pass
the context.

This change isn't documented in the change commit message and affects the
rollback, as now the context is passed always the compensation context
of the rollback initiating command is used opposed to before.
therefore i'm restoring the previous behavior here (null was passed as the
context for the callback).

Change-Id: I02cd0bf6ff89591f5f62741200b5620ac96fbd90
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1284250
Signed-off-by: Liron Aravot <laravot@redhat.com>