userportal: No up DC with cluster with permission message
It fixes missleading message when one tries to create a VM in userportal
without permissions for cluster. Previous message was: "Not available
when no Data Center is up." which might not necessar be true. New
message is "Not available when there is no Data Center up with Cluster
you have permissions for."

Change-Id: I0a3d5218a678f4d5862849698c90bbd0b035faa3
Bug-Url: https://bugzilla.redhat.com/1278271
Signed-off-by: Jakub Niedermertl <jniederm@redhat.com>