scheduler: Add in-cluster upgrade policy units
Add a filter policy to filter out all host which run an older OS than
the VM is running on or was running on the last time.

Add a weight policy to give older OS versions a very bad score, the same
OS version a not so bad score and newer versions the best score.

The filter and the weight policy unit are not part of any scheduling
policy by default.

Change-Id: Ifed268ae185b16d72bf207755f730bc55c0b1838
Bug-Url: https://bugzilla.redhat.com/1241149
Signed-off-by: Roman Mohr <rmohr@redhat.com>