infra/storage: send commands to the SPM on a FIFO basis
The synchronized 'syncObj' object, used to serialize SPM commands from
the manager to the hypervisor is not particularly fair, which makes
these commands to arrive in random order on vdsm. This is particularly
true when a lot of commands are sent to the manager, using the API for
example. In this case, some delays may get very large.

Using a fair ReentrentLock instead of executing the code in a
synchronized block, respects the order of the commands, eliminating the
chance of starvation.

Using a reentrant lock makes it easy to report the number of threads
waiting on the lock as well.

Change-Id: I23fb85a4179c8a77bf1b3df532d43edc4207f33a
Bug-Url: https://bugzilla.redhat.com/1277057
Bug-Url: https://bugzilla.redhat.com/1282764
Signed-off-by: Tim Speetjens <tim.speetjens@redhat.com>