restapi: Add model servlet
This patch adds a servlet that servers the representation of the model,
in its original source form and also in the XML and JSON
representations. The servlet is available in "/ovirt-engine/api/model".
It supports the following content types:

  application/json
  application/octet-stream
  application/xml

These content typs can be indicated using the HTTP "Accept" header, as
usual, and also the "accept" query parameter:

  GET /ovirt-engine/api/model?accept=application/json

File extensions can also be used to indicate the requested content type.
For example, the following request will indicate JSON:

  GET /ovirt-engine/api/model.json

Change-Id: I895a6923849402f8f392336c8aa69e03a8e90183
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>