core: PowerSavingBalancePolicyUnit streaming API usage
Replace the old-fashioned LinqUtils usage with the modern, built-in
streaming API introduced in Java 8.

This patch does not only replace an outdated syntax with a more modern
one, it also offers a (theoretical) performance improvement by using
findFirst's short-circuiting termination logic and stopping the
iteration once a host that fulfils the conditions is found.

This patch is part of the effort to remove LinqUtils completely.

Change-Id: I11ee8d61155c7e15b2a7a27ffe59872ad4d7fd4a
Signed-off-by: Allon Mureinik <amureini@redhat.com>