core: Add updateSnapshot to CinderDiskParameters.
Add a new parameter called updateSnapshot.
This parameter should indicate whether the snapshot should be updated or
not while removing the volume.
For example when we remove an image while restoring a volume the
snapshot should not get updated, but when we remove a disk we definitely
need to remove the volume from all the VM's snapshots.

Change-Id: I8af9e844d6c0513230ad0966d4547ac2584b3711
Bug-Url: https://bugzilla.redhat.com/1256048
Bug-Url: https://bugzilla.redhat.com/1256049
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>