core: profiles: make AddUserProfile update the key
To accomodate the need of REST API, we make AddUserProfileCommand
detect when the client wants just to update the SSH Public key
parameters, and we handle this case inside the command as special
case of Update, trasparently.

During the verification of this patch it was discovered that the
update command was not refreshing the SSHPublicKeyID after each change.
This is harmless for {UserPortal,Webadmin} UI, but incorrect for
the REST API.
So this patch also make sure that each succesfully update
changes the key ID.

Change-Id: Ia197bddb74186fa7c696f7e7f28e8fbe5e487e08
Bug-Url: https://bugzilla.redhat.com/1283499
Signed-off-by: Francesco Romani <fromani@redhat.com>