core: DeactivateStorageDomainCommand - refresh all hosts
Currently on DeactivateStorageDomain in a DC that uses the disk backend
we refresh the hosts that are used to move domains to Active status
while in dc that uses the memory backend we refresh also hosts in
statuses that we check before moving a domain to maintenance.
As currently the refresh operation always required an access to the
master domain, this patch unifies the behavior - the operation might
take long time if there's no access to the master domain, but on most
cases we'll benefit from it as domain will move to Maintenance status
faster if there is a non operational host.

Change-Id: I4c371e6188cdb93e1986996c2dc9b843889969bd
Bug-Url: https://bugzilla.redhat.com/1256841
Signed-off-by: Liron Aravot <laravot@redhat.com>