core: support for convergence schedule
Support for setting the migration policies.

The list of policies is set in the vdc_options under
MigrationPolicies. It is a JSON formatted string.
From the point of view of the user it is exposed only as a list of well
described policies.

The following fields are defined in this patch:
- convergenceItems: list of actions which will be executed as a
  reaction to stalling
- initialItems: list of actions executed before the migration starts
- lastItems: list of actions executed as last ones (e.g. the very last
  item
  should always be something which finishes the migration one way or
  the other (e.g. abort or turn to postcopy)

The maxDowntimeLimit can be overridden on the VM level
(custom migration downtime).

Change-Id: I389e202b81b01cc6891d6e5cea1a995525769987
Bug-Url: https://bugzilla.redhat.com/1252426
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>