webui: Fix numa pinning dialog cancel button
When opening the numa pinning dialog through the VM edit screen, the
pinning changes were persisted when you hit the cancel button. They
should only be persisted when the OK button is pressed.

Instead of directly manipulationg the numa nodes associated with the VM,
a numa node model is moved around. When hitting the OK button the model
is converted to numa node objects and set on the VM.

Bug-Url: https://bugzilla.redhat.com/1284784
Change-Id: Ifc7ac4cdb0f7887bf74de5cdba2ca35b7bc92dfc
Signed-off-by: Roman Mohr <rmohr@redhat.com>