packaging: setup: pki: Do not fail if pkcs12 unreadable
During upgrade, we check several pkcs12 files to see if they expired or
are close to expire. Do not fail upgrade if we failed to read them. This
can happen e.g. if some of them were replaced with ones signed by a 3rd
party CA, changed their passphrase, etc.

Change-Id: Ibe88e38b073433a357ee7ac09f6ead159e22baa4
Bug-Url: https://bugzilla.redhat.com/1286607
Signed-off-by: Yedidyah Bar David <didi@redhat.com>