restapi: Add the "async" parameter to all the actions
All the update, remove and action methods support an "async" parameter
that indicates if the operation should be performed asynchronously.
Currently this is documented in the RSDL, but not in the model. This
patch adds this parameter to the model, so that in the future it will be
available in the documentation.

Change-Id: I9e66d9ac3bc5b9549f4374d15d507ae063ee50f3
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>