core: Get initial Cinder volume to restore from.
Add a method which returns the initial Cinder volume the engine should
delete.
The restore process of the Cinder volume should use this volume to get
all its descendants and remove them all.

The initial volume is chosen by taking the previewed snapshot,
fetch the parent volume, and retun the other volume.

Change-Id: Iffaa62f703734f50208418786dc813bfe82106b9
Bug-Url: https://bugzilla.redhat.com/1256048
Bug-Url: https://bugzilla.redhat.com/1256049
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>