dbutils: Clean jobs when steps are cleaned
This patch insures that when the taskcleaner util deletes steps, the
corresponding jobs of those steps will be cleared as well.

Prior to this fix steps were cleared by taskcleaner while the related
job remains in STARTED status and actually without any steps.
This leads to Orphan jobs stuck in the system and not cleared forever.

Change-Id: I9c2f775a93208d667b11b4d32a93277f44d11776
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1283944
Signed-off-by: emesika <emesika@redhat.com>