core: Fix of repetitive setting same custom icon
Following scenario was previously failing:
 - use new custom icon (e.g. edit vm)
 - remove usage of custom icon (e.g. use default icon for vm)
 - upload the same custom icon again (e.g. edit vm)

The cause was that icon ids in parameters were sent from frontend
(because the icon was stored in frontend cache) but the ids were not
valid (icon was removed from db in second step).

This patch fixes it by ignoring icon ids in parameters by canDoAction
validation if actual icon is uploaded.

Change-Id: I72014b20cbe85b2a792cd9e8a53b4b65c641cf59
Bug-Url: https://bugzilla.redhat.com/1267711
Signed-off-by: Jakub Niedermertl <jniederm@redhat.com>