engine: Introudce NetworkAttachmentIpConfigurationValidator
The validator will validate the network attachment's ip configuration.

A valid ip configuration must has a primary address with valid
boot protocol and ip details.
In case the boot protocol is static, ip details must contains ip
addres and netmask.  Please note that gateway is not mandatory
as the kernal can complete it with a default one.
For dhcp or none boot protocols, no ip details should be
mentioned.  In case of dhcp, the dhcp itself should determine
it.

Change-Id: I589013ef2b6783c77a424a8d21fad5da8a7f3250
Bug-Url: https://bugzilla.redhat.com/1270807
Signed-off-by: Eliraz Levi <elevi@redhat.com>