restapi: Document the use of All-Content in /clusters
The "/clusters" resource uses the "All-Content" header to decide if
additional information should be displayed for each cluster. In
particular it uses it to decide if the "management_network" property
should be displayed. But the header isn't documented in the RSDL
metadata. As a result it isn't available in the SDKs or in the CLI. This
patch documents it.

Change-Id: I5cdf734ac9640a202006d46791a57d1d8ddc97c4
Bug-Url: https://bugzilla.redhat.com/1287688
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit f0091b8a246022b5f699268b83bcd41fb521c868)