core: Update executionContext with Steps.
When executing a command through the CommandExecutor as part of an
execution flow, the steps in the executionContext should be initialized as well.

At this particular bug a stateless VM is being powering up and a step indicating
the snapshot is being created is initialized in the context and should be passed to the
child command CreateCinderSnapshotCommand.

The proposed fix adds the initialization for steps and jobs to the
executionContext when calling createDefaultContextForTasksImpl

Change-Id: I4ee874e3da01504115e1fa2fd2e66de51d441a9d
Bug-Url: https://bugzilla.redhat.com/1270725
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>