core: Get previewed snapshot's images by the snapshot's OVF.
We should get images from previewed snapshot configuration, instead
fetching them from the DB.
The reason for that is to overcome a scenrio when a VM contains several disks,
gets previewed from a snapshot which only have part of those disks (Several
disks were created after the snapshot was created).

The preview operation, for the disks created after this snapshot, is to
remove their data from the base_disks table and only leave their data in
the images table.

The problem appears once the user undo/commit the previewed snapshot.
The functionality is adding the data to the base_disks and always assume
it is an image disk.

The proposed solution is to use the VM configuration when fetching the
disks to restore once a previewed VM gets undo/commited so the engine can
determined which type of disks we are restoring after an undo operation.

Change-Id: I421e95e3dd0023927f07db697dcb86b2593be848
Bug-Url: https://bugzilla.redhat.com/1256048
Bug-Url: https://bugzilla.redhat.com/1256049
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>