core: StorageDomainDaoTest cleanups backport
This patch backports a series of cleanups done to StorageDomainDaoTest
in order to facilitate bugfixes in the following patches.

It includes:
- Fixing naming conventions of the class' methods
- Removing a useless member
- Using constants from the FixturesTool
- Fixing the order the assertEquals parameters
- Removing redundant whitespaces that hinder the code's readability
- Removing some pointless assertions

Note this patch should also have a [very slight] performance
improvement, as it uses a pre-created Guids instead of invoking the
constructor several times and removing some useless assertions.

This patch contains the fixes introduced in patches 3158869..15d72f1
from the master branch.

Related-To: https://bugzilla.redhat.com/1337257
Related-To: https://bugzilla.redhat.com/1341661
Change-Id: I8f25dfbd3530b555a0c2aa3ff992990274be6db8
Signed-off-by: Allon Mureinik <amureini@redhat.com>