core: Remove ListUtils#firstMatch
Removed ListUtils#firstMatch and replaced it with calls to Java 8's new
streaming API as part of the attempt to remove ListUtils completely.

As an added bonus, this new implementation offers a slight performance
improvement by using a HashSet's contains method instead of performing a
binary search in a sort array.

Change-Id: I6eb9ba31753ecf4508dd2a5ed4f80236f6d67526
Signed-off-by: Allon Mureinik <amureini@redhat.com>