core: Unbreak DAO tests
Commit b3eacdb46f2b9c65ca184c9352b39f766d0c936b introduced Java 8
bytecode to the Ovf*Reader classes. While these classes are not used by
any way, shape or form by the DAOs, they are part of the utils project,
which the dao project depends on. Thus, when the DAO tests scan for
beans, they encounter this class, and fail since Spring 3 we are using
cannot handle Java 8 bytecode.

This patch fixes the tests by moving the CdiIntegration test class to a
sensible place (org.ovirt.engine.core.dal), and reducing the scope beans
are searched for to org.ovirt.engine.core.dal and
org.ovirt.engine.core.dao.
As a side bonus, this patch offers a slight performance boost to the
tests, as they no longer need to scan the entire org.ovirt.engine.core
huge package.

Change-Id: I8c7a749d5c37f4dcbfd95cea1e01a93520d2ce87
Signed-off-by: Allon Mureinik <amureini@redhat.com>