core: Remove ListUtils.getAddedElements
ListUtils.getAddedElements has the same functionality as the standard
org.apache.commons.collections.CollectionUtils.subtract, and thus can
easily be removed for sake of reducing the size of the codebase and
focusing on oVirt's business logic.

In the only place it was used, it was used to subtract between two sets
with a maximal size of 2, so performance is hardly an issue here.

Change-Id: Ia7a0bf9c1f897fca7eafeb279bc41776c794526f
Signed-off-by: Allon Mureinik <amureini@redhat.com>