core: before unlock a VM, fetch it from the DB
The VM which is used to be unlocked is the VM from the meory, so the VM
might be with DOWN status although it was locked on the proccess.

Since CoCo handle the class in the memory (unlike the Async Task
functionality), after commiting/undo a previewed Cinder snapshot, the VM
from the memory is being unlocked but since the status of the memory of
the VM is DOWN, the engine doesn't unlock it.

The proposed fix is to make the VM property to be null in order to
refresh the VM status from db.

Bug-Url: https://bugzilla.redhat.com/1256048
Bug-Url: https://bugzilla.redhat.com/1256049
Change-Id: I7edd4e8b23277c57939bd5e965eb4da29e3600d3
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>