webadmin: Handle ClassCastException in import VM/template popups
When importing a VM or template, before there was the possibility to select
a destination storage domain per disk we had a default storage domain
selection, this is not used or visible for quite some time now and aside
for throwing ClassCastException this model did nothing.
Up till recently, UI exceptions did not propagate to the client and was
ignored, now when UI exceptions are logged at the client it is visible to
the user, this patch removes the remnant of this redundant model thus
preventing the ClassCastException.

Bug-Url: https://bugzilla.redhat.com/1289037
Change-Id: I9c1b69b81e01bd82265927fefb4e781969f15b14
Signed-off-by: Tal Nisan <tnisan@redhat.com>