restapi: "id" as XML attribute
Currently some of the entities of the model explicity define an "id"
attribute instead of extending the "Identified" interface. For example,
the "LogicalUnit" entity defines its own "id" attribute. As a result the
generated XML complex type contains an inner "id" element:

  <xs:complexType name="LogicalUnit">
    <xs:sequence>
      <xs:element name="id" type="xs:string" .../>
      ...
    </xs:sequence>
  </xs:complexType>

But the "id" attribute should always be represented as an XML attribute,
as that is what was done in previous versions of the API, and it is a
common practice in other usages of XML.

To solve this problem this patch changes the XML schema generator so
that it will take into account that some model attributes need to be
represented as XML attributes instead of inner elements. The "id"
attribute is the only one that needs this special treatment at the
moment, but the pach implements a more generic approach, just in case.

The resulting representation, for example for "LogicalUnit", will be the
following:

  <xs:complexType name="LogicalUnit">
    <xs:sequence>
      ...
    </xs:sequence>
    <xs:attribute name="id" type="xs:string"/>
  </xs:complexType>

Change-Id: I046b8bd2f71e0add754e4193885aeb8ba7110e70
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>