core: Add InClusterUpgrade policy
Host OS version mixing is now only possible if the config flag
ConfigValues.CheckMixedRhelVersions is set to false AND afterwards the
InClusterUpgrade policy is enabled.

So the config value---from now on---just enables the possibility to set
the provided InClusterUpgrade policy. When the policy was set
successfully, adding host with different major OS versions is possible.

Change-Id: I5acf7be5f608698b87091c10d8364e5973d3b441
Bug-Url: https://bugzilla.redhat.com/1241149
Signed-off-by: Roman Mohr <rmohr@redhat.com>