core: DiskImage.hasRawBlock()
isRawBlock() is used to check whether a DiskImage needs refreshing on
VDSM's side before an action is performed on it.

Theoretically, a DiskImage can have several copies on several domains,
and only some of them may be block domains (thus requiring a refresh).
This patch removes the assumption that the block domain would
necessarily be the first domain returned in getStorageTypes() and
instead checks all the domains to see if at least one of them is a block
domain.

Currently, this situation is only possible for templates, which don't
participate in Live Merging, so this patch is more of a future-proofing
effort than an actual bugfix.

Change-Id: Ief879683c433e52efcf4c2d2dc43db81f0475835
Signed-off-by: Allon Mureinik <amureini@redhat.com>