core: ignore noVm error for internal destroy ops
In Icd93048d78 we changed the engine to ignore errors of type noVm
when they arrive from the destination host after cancel migration.
Actually this is a particular instance of a more general problem -
if we ask to destroy the VM and it is not there anymore, does it
really mean that the opreation fails?

Sure, for user-initiated operations it makes some sense to report
these errors so users will know that the operation they asked for
failed (and this is also debateable), but for internal operations
- it is not interesting for anyone to know about this "problem".

Therefore, setting ignoreNoVm to true for all destroy operations
that are done from the monitoring.

Change-Id: Ibc43d294804b0dd4070c0a08af6bba40652e1281
Bug-Url: https://bugzilla.redhat.com/1290361
Signed-off-by: Arik Hadas <ahadas@redhat.com>