restapi: Optimization of DB access in REST GET vms
Number of DB queries for getting graphics devices within
a REST GET VMS List call is constant (2).
It was linear with the count of VMs before.

New GetGraphicsDevicesMultipleQuery is added.
The GetVmDeviceByTypeAndDevice stored procedure supports filtering
and the result set is filtered on the given UUID[] vm_ids list.

Index on vm_device 'type,device,alias' is added.

Change-Id: I9040c0790f5b10860fbdfed6ba6e1667e5901d83
Bug-Url: https://bugzilla.redhat.com/1268216
Signed-off-by: Marek Libra <mlibra@redhat.com>