core: Image Upload backend commands
Image Upload allows a client to transfer a VM disk image into the oVirt
environment.  The backend commands are responsible for creating and
managing the disk images, manging sessions and ticketing between the
client, proxy, and host, and performing any cleanup upon completion or
failure.

The following new commands are added:

UploadDiskImageCommand is the parent command that creates a disk image
for VM disk uploads.

UploadImageCommand is the base class responsible for managing sessions
and tickets as well as coordinating the image transfer between the
various involved components (proxy, vdsm, vdms-imaged, ui).

UploadImageStatusCommand is used by clients to update and retrieve the
status of an ongoing UploadImageCommand.

Change-Id: I8244824f33e3d2b5799bf9c01114e38e1d162583
Related-To: https://bugzilla.redhat.com/1091377
Signed-off-by: Greg Padgett <gpadgett@redhat.com>