frontend: don't auto format sd when detaching from local dc
This patch adds the user the ability to choose whether or not to format
the storage domain when detaching it from a local data center. This is
done by presenting him a checkbox with the label "Format Domain, i.e.
Storage Content will be lost!", like in the "Remove Storage(s)" window.

Change-Id: I78526987f2b6df1e76423c2b74b0d0a5be26032e
Bug-Url: https://bugzilla.redhat.com/1245249
Signed-off-by: Idan Shaby <ishaby@redhat.com>