core: Remove VM#runOnce
The private field VM#runOnce is never set, and is only read in the
hashCode() method. Hence, it could savely be removed.

It's worth noting that a VM actually does have a runOnce property
exposed by the isRunOnce() and setRunOnce(boolean) methods, but it's
value is stored by the VmDynamic object the VM holds.

Change-Id: Icbf21832633d0b0ff86abde7af3fae4388c8180b
Signed-off-by: Allon Mureinik <amureini@redhat.com>