engine: Limited use of CPU Profiles by user
Adding/Updating VM using REST API now checks that the user has
permissions for the CPU Profile of the VM.
This applies whether the CPU profile is changed or not. It
means that even if the CPU Profile is not changed or the user tries to
create a VM in a cluster where he/she doesn't have permissions for any
CPU Profile, he/she will not be able to Create/Update the VM.

This change effects the following Commands in the engine:
AddVmCommand, UpdateVmCommand.

In case no Cpu Profile was specified in the REST API request, Any CPU
Profile that the user has permissions for will be used to comply with
the old behaviour of the system

Change-Id: Ic187e1815fbb2334e4eca7140e6b2a5493059bae
Bug-Url: https://bugzilla.redhat.com/1143869
Signed-off-by: Tomer Saban <tsaban@redhat.com>