core: compare dedicated hosts against template on add vm
On add-VM we used to require permission on action-group
ActionGroup.EDIT_ADMIN_VM_PROPERTIES if the VM we get in the parameters
contains non-empty list of dedicated-hosts. By default, this list
contains the dedicated hosts that are set in the template. So as a
result we required users to have that permission in order to create a
VM from any temaplte that is set with dedicated hosts, even when they
did not change the template's setting.

This patch fixes it by changeing the check in add-VM so it will compare
the list in the given VM with the list that is set in the template so
ActionGroup.EDIT_ADMIN_VM_PROPERTIES will be required only if the list
of the dedicated hosts was changed by the user.

Change-Id: Iab1a8536e1da5c529d735542d21b004a15c5fbf4
Bug-Url: https://bugzilla.redhat.com/1284472
Signed-off-by: Arik Hadas <ahadas@redhat.com>