core: hosted-engine: import the domain only when DC is UP
Most of the setups will be concluded by an engine restart
(configurable). At that stage the import may have already kicked in, and
this race causes -
1) a reboot during the import - may lead to failed
   rollback, so we're left with junk in DB
2) An attemt to Attach before we
   have an UP host. This leads to a rollback which may lead to closure 1.

As long as we don't attempt to add a data domain during HE setup, this
approch will keep us safe and will import the domain while there isn't a
restart planned.

Change-Id: I2710be830cdb234fbb38456234219389e9299833
Bug-Url: https://bugzilla.redhat.com/1293928
Signed-off-by: Roy Golan <rgolan@redhat.com>