core: change vm name on import from vmware
The import dialog allows the user to specify a new name for a VM that
is imported from VMware. However, it would never work since we send
this name to v2v as the identifier of the VM to import - so most
probably it won't work since there will be no VM in VMware with that
name. In the worse case, we might import the wrong VM..

Change-Id: Iea7f45a5d016352e44807dc85042eb717a170309
Bug-Url: https://bugzilla.redhat.com/1294651
Signed-off-by: Arik Hadas <ahadas@redhat.com>