core: Added permissions on AddProfileBase
When adding Disk/Cpu Profile there is always a need to add permissions
on the profile. For that reason, I added an abstract method
addPermissions() that will run each time add profile command is
executed. This method will be implemented by the add Disk/CPU profile
commands with the necessary permissions.

The developer may cancel the automatic call of addPermissions by giving
false in the constructor.

Also, Create/Update/Remove and 'Assign' permissions are given to the
user that created the profile and 'Assign' permissions to everyone else.

Also, this patch changes the way the CPU Profile is added when running
the 'AddClusterCommand'.  Instead of adding the CPU Profile directly to
the Database which bypasses all the necessary validations, this patch
uses the 'AddCpuProfileCommand'.

Change-Id: I6f65daf1e46be9521e8541932ce84710175a6c66
Bug-Url: https://bugzilla.redhat.com/1143869
Signed-off-by: Tomer Saban <tsaban@redhat.com>