core: Filter out child Cinder commands on end action.
EndAction was being called only for image commands.
Cinder commands should get filtered out on endAction since the logic
is already being done as part of the Cinder command endAction.

In the future this filtering should be removed and the child endAction
should be called abstractically for Cinder and Image volumes.

Change-Id: Ia8e5119ec55f498b4a688334cd6302abe7287388
Bug-Url: https://bugzilla.redhat.com/1293226
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>