restapi: Initial support for multiple versions
This patch adds initial support for having multiple versions of the API
in the same server. It covers only the metadata part, that is, the XML
schema and the RSDL.

The versions supported will be "3" and "4".

The metadata for version "3" is just a copy of the metadata files used
in the 3.x branch of the engine. This will need to be updated from time
to time to keep it in sync with the 3.6 branch, till that branch is
definitively closed.

The metadata for version "4" is automatically generated from the model.

The patch also adds initial support for the "Version" header, so that
the user can indicate what version of the API to use, right now only for
the metadata. For example, to request version 3 of the XML schema the
user can send a request like this:

  GET /ovirt-engine/api?schema
  Version: 3

Same for the RSDL:

  GET /ovirt-engine/api?rsdl
  Version: 3

When the "Version" header isn't provided the default will be used,
whichis currently version 4.

Change-Id: I13adfd3feb0a3d56133facb399a93c2e7e0d9ac3
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>