core: Add force compensation when creating a VM from Template
Adding a VM from Template and restart the engine might cause the VM to
be locked forever in the DB once the engine is started again.
The reason for that is because the compensation is not enforced in the
AddVmCommand,

Adding forceCompensation=True to the command's annotation should solve
this and remove the VM form the DB once the engine gets restarted.

Change-Id: Ibc465fe9dbd36e7582583865d185e35844cf3b76
Bug-Url: https://bugzilla.redhat.com/1297190
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>