core: stop upgrade when shell script fail
This patch fixes a bug that occurs when we have a shell upgrade script
that fails to execute.
Current behaviour is that the script fails and upgrade process continues
and even generates a SUCCESS result on CI tests.

This patch adds a missing check on the result of running the script and
terminates upgrade procedure if shell script fails with a proper
message.

Change-Id: Ic18a8a567945c4d3e709d28b85691f0e3b67e2e5
Signed-off-by: emesika <emesika@redhat.com>