frontend: fix FcpStorageView faults
FcpStorageView has four faults that are fixed in this patch:
1. It extends from AbstractSanStorageView, which is
useless since it's the only subclass of AbstractSanStorageView.
2. The corresponding ui.xml file is AbstractSanStorageView.ui.xml (and
not FcpStorageView.ui.xml), which causes the editor to ignore the label
"errorMessage" in FcpStorageView and to malfunction.
3. The @Path annotation of the label "errorMessage" is "GetLUNsFailure",
while the corresponding string field in the model, SanStorageModel, is
"getLUNsFailure".
4. AbstractSanStorageView extends from
AbstractStorageView<SanStorageModelBase>, while the model of
FcpStorageView is SanStorageModel.

This patch:
- Merges AbstractSanStorageView into FcpStorageView.
- Renames the AbstractSanStorageView.ui.xml to FcpStorageView.ui.xml.
- Renames the Path annotation of the label "errorMessage" to
"getLUNsFailure".
- Replaces the generics type of AbstractStorageView to SanStorageModel.

As a consequence, the label "errorMessage" is finally shown when it's
supposed to be shown in the FcpStorageView.

Change-Id: Idff765f50ce3d461c18f0db428687e7a27e8b64c
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1297689
Signed-off-by: Idan Shaby <ishaby@redhat.com>