core: fixing a problem in upgrade
The last rename of "vds group" to cluster caused a problem in schema.py
on upgrade, at that point, the script looking for cluster versions does
not have the cluster table and need to use vds_groups.
Solution was adding in an upgrade script a view named vds_groups that
only selects the original cluster table records and revert schema.py to
relate to vds_groups.
This is done in an upgrade script and not in create_views.sql since
eventually when this code will be squashed to the base DB, we will drop
this vds_groups view.

The reason why this error was not found by CI tests is that upgrade jobs
are using schema.sh script directly and not engine-setup which seems to
be problematic.

Change-Id: If230d3cd397c314a27cc78d3837b637b671ba706
Signed-off-by: emesika <emesika@redhat.com>