core: Update the OVF store when hosted engine VM is edited
Retrigger the ovf update from the UpdateVmCommand when
the hosted engine VM was edited.

This is needed to persist the hosted engine VM configuration
as soon as possible in case the ovirt-hosted-engine-agent
needs to restart the engine VM with the proper settings.

There is still a very slight chance of hitting a race
condition (when the regularly scheduled OVF updater processes
the hosted engine VM at the exact same moment as the triggered
job), but the VM will be saved during the next OVF update anyway.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1343991
Change-Id: I121a9f587b2a85bdd4b366f07b040dc85fa515cf
Signed-off-by: Martin Sivak <msivak@redhat.com>
(cherry picked from commit 7e92bd44e17ded6831eceee1610e1b3616c9c505)