webadmin: auto-scroll SetupNetworks interfaces list during drag-and-drop
When a host has more the 6 network interfaces, the Setup Host Networks
popup dialog needs to scroll to access all of the interfaces.  During
a drag-and-drop of a logical network or network label, the interface
list did not allow scrolling.  This patch adds auto-scroll handling so
a user can access all of the host's available interfaces.

To auto-scroll the interfaces list:
  1. start a drag operation on a logical network, label or
     interface
  2. move the mouse pointer close to the top of the interface list
     to auto scroll the list down
  3. move the mouse pointer close to the bottom of the interface list
     to auto scroll the list up

Scroll will stop when the drag operation is complete or the mouse
is moved outside the autoscroll zone.

Change-Id: I2c6fe977f014309659d5559d4609d9361a43274e
Bug-Url: https://bugzilla.redhat.com/1249895
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>