engine: Do not assign creator as owner to VMs from pool
If VM is created to attach it to a VM pool, do not assign its creator as
owner in any case. Assignment of a user to the VM in this case will
designate this VM as occupied and prevent other users from taking it.
This assignment happens currently when pool creator doesn't have
MANIPULATE_PERMISSIONS role on the pool VMs.

Change-Id: I6af18a2b7caa8181523a277f4d63cc688e69a59e
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1301867
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>