backend: create a single API for command creation
Currently, creating a command for the positive usecase (i.e., using a
parameters object, as opposed to the compensation usecase when the
backend bean is loaded) can be done using two different constructors:
one that gets the parameters object, and the other that gets both the
parameters object and the command's context.

Since we use reflection to create the command, this approach has several
faults:
1. Even though a non-null CommandContext is passed to
   CommandsFactory.instantiateCommand, there's no way of knowing
   a-priori whether the overloaded constructor with a CommandContext
   parameter exists. Hence, two calls will be executed in order to find
   it - first, hasConstructor is called in order to determine what
   constructor the command class has, and only then
   findCommandConstructor is called with the appropriate arguments.

2. There might be dozens of constructors that are never used, which
   means a lot of confusing dead code (will be cleaned in the next
   patch in the series).

3. The worst - if createCommand is called with a CommandContext
   argument but the command lacks the appropriate
   constructor, a default CommandContext will be created for the command
   and the instantiateCommand call will then apply ONLY the
   ExecutionContext from the passed argument to the command. This might
   lead to a hoard of elusive bugs as the caller thinks he passes an
   entire CommandContext while most of its data (such as locks, the
   compensation context, etc.) will be silently ignored.

This patch aims to rectify the aforementioned situation, by forcing
the commands to have only one mandatory constructor (which receives
both a parameters object and a context) for the positive usecase.

This patch contains the following changes:
* Adding the mandatory constructor to any command that doesn't already
  contain it.
* Removing the unused constructor from all the commands (the one that
  gets only the parameters object).
* Adding a test to enforce this behavior.
* Simplifying the commands' creation to use only the mandatory
  constructor.

Note that this doesn't mean that a command can't have an additional
constructor that only receives the parameters object, e.g., for using
in unit tests. It just means that the mandatory constructor will have
to exist since it's the only way to instantiate the command via the
CommandsFactory.

Also note that CommandsFactory's public API was not changed, and you could
still use createCommand(VdcActionType action, P parameters). Only the
implementation was changed.

Change-Id: I8a2747fb23fc9e66910642d806d14518949d6d4d
Signed-off-by: Idan Shaby <ishaby@redhat.com>