core: Refactored ProfileParametersBase constructor
ProfileParametersBase accepted both Profile and Id. Id can be retrieved
from profile.getId() instead of as an argument in the constructor.
This is bad practice and has potential for problems. So, now the
constructor only accepts a profile and the id is taken from the profile
itself.

Methods that were using the constructor were changed inorder to comply
with the new constructor arguments.

Change-Id: I0b9fa9ff81422fa5a9782e847892d79f2e5e7b74
Bug-Url: https://bugzilla.redhat.com/1143869
Signed-off-by: Tomer Saban <tsaban@redhat.com>