core: adding missing ctors
When the AsyncTaskManager creats a command it creates it without
the context.
When migrating the different flows to use coco rather than rely on
the async task manager, the modified commands that were used as child
commands until now will be created by the async task manager.

This patch adds the missing parameters ctor to those commands.

Change-Id: Iaae93b7e6b041749c4911806bfdb78fa4afe166c
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1300757
Signed-off-by: Liron Aravot <laravot@redhat.com>