webadmin: add arbiter support for gluster volume
 Adding support for arbiter volume creation. Arbiter Volume
check box will be added in new volume dialog and add brick
dialog and it will be shown when user creates a replica volume.
Arbiter volume should be a replica 3 volume and every third brick
in the brick list will be treated as arbiter brick.

 Arbiter Count should be 1. Hence it is shown as a check box
in UI and converted to 1 or 0 during volume creation.

Change-Id: I7d24653a633b0ccc6b6676af24b985bdd3a0be8a
Signed-off-by: Ramesh Nachimuthu <rnachimu@redhat.com>