core: NetworkAttachment should use AnonymousHostNetworkQos
this patch alters NetworkAttachment not to use HostNetworkQos for
overriding HostNetworkQos values. This was former approach and there
was problem with bean validation, which failed if name/DC id was not
provided in that override.

This patch solves ONLY bll layer. Db
and rest layer is still the same.

ad DB layer: all qos are in QOS table, without referential integrity
check, allowing null valued name and storage_pool_id. Although
unfortunate design, this does not pose any current problems.

ad REST layer: with changes in this patch it will work, however OO
design is not OK, as overriding qos values is not the same as
HostNetworkQos — user still might pass name or DC id and those values
will be silently ignored

Change-Id: I8d5fcce039ee5c24ece32cec006a85db79757949
Bug-Url: https://bugzilla.redhat.com/1329224
Signed-off-by: Martin Mucha <mmucha@redhat.com>