restapi: Use JAX-RS 2.0 API artifacts
The engine has been recently updated to work with WildFly 10. This
version of WildFly implements JAX-RS 2.0, but we are still using 1.x
artifacts for the build. This doesn't generate a real problem, but it
may in the future. This patch changes the POM files so that they use
versions of the artifact matching those provided by the application
server.

In addition, we are currently explicitly using the JAX-RS API artifacts
provided by Resteasy, which introduces a bit of locking. To avoid that
this patch changes the POM files so that they use more implementation
independent artifacts where possible. The Resteasy artifacts still be
needed by the tests, as they need an actual implementation of the JAX-RS
API.

Change-Id: I97ae0d6c9762d7e2130f4a1780a85bf8c1de38ec
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>