core: User version 3.1 of the servlet API
The engine has been recently updated to use WildFly 10, which uses
version 3.1 of the servlet API. But we are still building using version
3.0. This doesn't currently create any problem, but it is good practice
to build with the same version that will be used during runtime. This
patch changes the POM files so that the version used for the build will
be the same used during runtime.

Change-Id: I15f299a5ecbf462305785ec23aae0527ac76c2d5
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>