core: AddDiskCommand - CoCo usage when adding image disk
This change adjusts the AddDiskCommand to use the CoCo infrastrature
when used to add an image disk.
The command will use a callback to track its child commands on two
scenarios -
a. Its parent command has a callback
b. There is no parent command

When all the parent commands that executes AddDisk will use CoCo
and the add cinder disk flow will be adjusted we'll set the command
to use callback always.

Change-Id: If5943654ff25144b967be1ef3271727e663dc6cd
Signed-off-by: Liron Aravot <laravot@redhat.com>