frontend: Fix name uniqueness chack in ImportVmModel
Now when checking the name uniqueness of imported VM, we check for conflict
only VMs in target data center.

To facilitate this the data center loading part from
ImportVmFromExportDomainModel#doInit() and
ImportVmFromExternalProviderModel#doInit() had to be extracted into special
helpers withDataCenterLoaded() that are used to wrap the setItems() method doing
the name uniqueness analysis - since it requires access to storagePool.id.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=912263
Change-Id: I86c8b17bee0d676bc09c758e98439e5d672eb885
Signed-off-by: Martin Betak <mbetak@redhat.com>