restapi: Improve multiple version support
This patch adds a new HTTP filter that is responsible for extracting the
version information from the request, and adding it to the object that
handles the information about the current request. This used to be part
of the "current" filter, but it is better to separate the two concerns.
The new filter supports extracting the version from the HTTP "Version"
request header:

  GET /api/...
  Version: 4

Or from an optional prefix of the request URL:

  GET /api/v4/...

Later patches will use this version information to dispatch requests to
different versions of the API.

Change-Id: I32881539b7489f51109b03e34b4713263221b288
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>