restapi: Simplify generation of root service links
Currently the links generated for the root service (for the /api URL)
are calculated using the JAX-RS "UriInfo" class. In order to simplify
these calculations, and in order to be able to support multiple versions
of the API, this patch changes the root service so that it uses instead
the methods provided by the "current" object.

Change-Id: I9de13aa71e0587bb31add94e7609185082f19be9
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>