restapi: Split link helper
Currently the "LinkHelper" class contains method that work with links
that don't depend on the current request and also with links that do
depend. The methods that work with links that dpened on the current
request will in the future be changed to use the "Current" class instead
of the "UriInfo" class, and to do so the class needs to be part of the
"jaxrs" module. To enable that future change this patch splits the
"LinkHelper" class into two: "LinkHelper" and "LinkCreator". The first
will contain the methods that work with the current request, and the
second will contain the methods that work with links that don't work
with the current request.

Change-Id: I61c48887e8ba4bf5fd23905d51e24d95dce8f5ca
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>