core: Added cpu profiles permissions propagation
Cpu Profile permissions will now propagate from parent objects of Cpu
Profile.
For example, If you have CpuProfileOperator permissions on a cluster
that has Cpu Profile 'A' then you will automatically have
CpuProfileOperator on 'A' and all the other Cpu Profiles on that cluster.

The list of objects that Cpu Profile permissions progpagates from is:
System -> Storage Domain -> Cluster -> Cpu Profile.

This also solves the problem that admin user (That has SuperUser
permissions on System) couldn't run any virtual machines because it
needed to have specific permissions for a CpuProfile on it's cluster.

Change-Id: Iaf5c142f8e563377cc262bf36ea5a44060b90d81
Bug-Url: https://bugzilla.redhat.com/1143869
Signed-off-by: Tomer Saban <tsaban@redhat.com>