core: extract unmanaged hosted engine code from vms monitoring
Extract the code that handles installed hosted-engine VM from VmsMonitoring
to simplify VmsMonitoring and reduce duplicated code with AddUnmanagedVmsCommand.

The analyzer treats installed hosted-engine VM as an external VMs and pass it
to the AddUnamangedVmsCommand. The command in turn identifies installed hosted-
engine VM as external hosted-engine VM that should be imported.

A positive side-effect of this change, besides the ones mentioned above, is
that we now store the static part of the VM before flushing the refreshed data
to the database and therefore at the end of the first cycle in which we get the
VM we will have the VM presented with the right status.

Change-Id: Ie19247de8093bd79e8bb3ae62233d79853bbb601
Signed-off-by: Arik Hadas <ahadas@redhat.com>