backend: remove redundant test testInternalAnnotationCtors
CommandCtorsTest.testInternalAnnotationCtors verifies that all the
commands that are annotated with @InternalCommandAttribute contain a
constructor which its last argument is CommandContext (only those who
are listed in VdcActionType, i.e, not abstract command classes).

Since testCommandMandatoryConstructorsExistence verifies that all these
commands contain the mandatory constructor (a constructor that
receives parameters and context objects), it also covers the validation
of testInternalAnnotationCtors.

Change-Id: Iab7899fb5d18c9ca79e2276734e371885a41f86d
Signed-off-by: Idan Shaby <ishaby@redhat.com>