core: StorageDomainDaoTest assertNulls
executeReadList's contract specifies that when a method that returns a
List has no data it returns an empty list, not a null. Thus, asserting
this is wrong on several levels:
1. It represents a break in the abstraction - a unit test should only
   assert the behavior of the class its testing, not other classes.
2. It's pointless, as the contents of the list are always checked
   immediately afterwards. If the list were null, the test would just
   error out with a NullPointerException.

This patch removes those aforementioned useless asserts. Once they are
removed, the assertCorrectGetAllResult becomes a wrapper to asserting a
list is not empty, so it was just inlined.

Change-Id: I84ea47545ed9f4bf2375187843906da942c92d7e
Signed-off-by: Allon Mureinik <amureini@redhat.com>