core: Improve calculation of disk actual/committed size
The various views containing the disk actual and committed sizes on
storage domains was done by functions which performed aggregate queries,
per storage domain.

By converting them to joins, the calculation can be done in parallel on
all the domains, like databases are supposed to be used.

This patch performs the following:
1. Creates the storage_domains_image_sizes view to encapsulate the logic
   previously performed by the fn_get_disk_commited_value_by_storage and
   fn_get_actual_images_size_by_storage stored functions.
2. Remove the aforementioned functions.
3. Add some assertions to the DAO test to ensure the functionality is
   preserved.

Change-Id: I1c76870153d55c4a4259ecaf6e9b96297f16aa72
Related-To: https://bugzilla.redhat.com/1302794
Signed-off-by: Allon Mureinik <amureini@redhat.com>