packaging: Never repack .jar files
Currently we conditionally disable repackaging of .jar files, setting
the "__jar_repack" global macro to 0 only for quick builds. But
repackaing isn't really useful for us. See a description of the reason
behind repackaging here:

  https://www.redhat.com/archives/fedora-devel-java-list/2008-September/msg00041.html

This patch changes the .spec so that the "__jar_repack" macro is always
set to 0, thus speeding up all the builds.

Change-Id: I4fcf6e7bcd0209fb9715539c95419a634217069a
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>