core, db: View, DAO and entity updates for Image Upload frontend
The progress of an image upload operation will be displayed in the
webadmin disk list.  In addition, the status will be used to determine
which upload operations are available (upload, pause, resume, etc).

In order to do this efficiently, the status of the upload should be
made available to the UI via propagation through the Disk object.

Change-Id: I3f9c17839086a2b63c65246a785b53cbe2829885
Related-To: https://bugzilla.redhat.com/1091377
Signed-off-by: Greg Padgett <gpadgett@redhat.com>