restapi: Add V3 capabilites for version 4 of the server
There are users of the API that get the version number of the product
(from the API itself, or from other places) and then expect to find a
capabilities description for that version. But these descriptions of the
API have been copied from version 3.6 of the server, which doesn't
contain a description for 4.0. This means that those user will fail. To
avoid that problem this patch modifies the V3 capabilities server so
that it takes the capabilities of version 3.6 and duplicates, but
setting the version number to 4.0. This way those users will continue
working.

Change-Id: I627eee5db1e79634ef83a76101139966550ec446
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>