frontend: Send "Version" header to the API
In the near future the server will support multiple versions of the API,
in particular it will support versions 3 and 4. The users of the API
will be encouraged to explicitly indicate what version they want to use.
To do so, the will be asked to send the "Version" with every request.
The value of this header will be the requested version. Alternatively it
will also be possible to specify the requested version as a prefix in
the URL, like "/v3/vms" or "/v4/vms". When the version isn't explicitly
specified the server will select a default, which will be configurable
in the future (currently it is "4", and it is hard-coded). To avoid
surprises if the default version is changed this patch modifies the
frontend so that it will always explicitly request version 4 of the API,
using the "Version" header.

Change-Id: I18f93fd451639f773aa61ec44ff5e53e0ce5e146
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>