core: serialization/deserialization of SnapshotType
Because of the toString() implementation of SnapshotType (which is used
in returned messages) its deserialization is broken.
This patch fixes the serialization issue, as this code seems unneeded
to me I've removed it in change I8fb722dd15f0ccc2ac058765bfd6802a2a21134c.

This patch just fixes the serialization is it should be backported to the
3.6 branch and I'd like to minimize the changes and testing scope there.

Change-Id: Iee7192bd4e3bdf0b56ee57d4b25c5fcc17190be7
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1303840
Signed-off-by: Liron Aravot <laravot@redhat.com>