core: Removal of SnapshotType name and I18n improvement
Currently the ACTION_TYPE_FAILED_VM_SNAPSHOT_TYPE_NOT_ALLOWED
message supports multiple snapshot types set generically to
the message.

In practice, this message is used with one status only (Regular)
and the generification requires specific serialization/deserialization
code in the SnapshotType class and lacks i18n support.

As this message is used with that status only, this patch updates it
to specify it explictly (thus making it i18n compatible), if we'll need
to change add support to more types in the future that can be done, right
now we can remove that and the related code.

Change-Id: I8fb722dd15f0ccc2ac058765bfd6802a2a21134c
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1304653
Signed-off-by: Liron Aravot <laravot@redhat.com>