restapi: V3 compatibility for VNIC profiles
In version 3 of the API it wasn't mandatory to specify a VNIC profile
when adding a NIC, instead the "network" and "port_mirroring" elements
were used to specify the network that the NIC should be connected to and
the port mirroring configuration. This support was removed when version
4 of the API was introduced. Now that the API supports multiple versions
this can be restored, without touching the backend or the implementation
of version 4, just modifying the V3 adapters and servers.

Change-Id: Ibc6bdfb7cdce5c85f315911ea488695b56553c9c
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>