core: Propagate lock state when running async commands via CoCo
A command may need to propagate acquired locks to executed child
commands, such as when AddDiskCommand calls AddImageFromScratchCommand.

For Image Upload, I had temporarily changed the above call to use the
Command Coordinator and discovered that the lock held by the parent
was not being passed to the child.

The root cause was the lock was being overwritten in CommandBase when
the execution context for the child was initialized.  To avoid this,
we now pass the lock defined in the Command instance to the context
inititialization method.

Change-Id: I8cab6aa4d8788703285fb43003995f7f9d7b4efb
Related-To: https://bugzilla.redhat.com/1091377
Signed-off-by: Greg Padgett <gpadgett@redhat.com>