core: Remove TagsDirectorTest#testCloneTag()
The Tags class has no special logic or implementation of cloning or
serializing. The testCloneTag() uses
apache.commons.collections.Trasformer to clone a Tags instance, but no
production code does this, meaning this is essentially a test for
apache, not oVirt Engine's code.

In short, this test is completely redundant, and hence this patch just
removes it.

Change-Id: I665e44aa1e004cbd9ffd66c1992d414187fea3fd
Signed-off-by: Allon Mureinik <amureini@redhat.com>