restapi: Add null servlet
When there is nothing mapped explicity (without patterns) to the root of
the application, the servlet container assumes that requests for this
root without a trailing slash correspond to a directory, and as a result
it redirects, using the HTTP 302 code, to the same root with a trailing
slash. For example, if the client sends this request:

  GET /ovirt-engine/api HTTP/1.1

The server will responds like this:

  HTTP/1.1 302 Found
  Location: https://engine.example.com/ovirt-engine/api/

This works fine for browsers, but it isn't supported by the SDKs, as
they don't understand HTTP redirections.

To avoid this issue this patch introduces a "null" servlet, that does
nothing and is explicitly mapped to the root of the application.

Change-Id: I45ddcc86b519cca6422f9347eaa18a18b305df53
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>