core: cache host's numa nodes
Analyze of profiling dumps shows that a significant portion
of the CPU time is spent on getting the numa nodes on the host.
This can be improved by querying them only once in the host
polling instead of doing that per-VM.

We can cache it because it is not expected to change so it will
be queried once and not per-polling operation, but it would
maybe make sense to query that nodes along with the VM numa
nodes in one query per-polling operation.

So for now the optimization is just to reduce the frequency
of the query of the host's numa nodes to be once per-polling.

Change-Id: I5d72ab207f861ca8e050068d101c998343cffc60
Signed-off-by: Arik Hadas <ahadas@redhat.com>