core: pass dependencies of vm-analyzer using the ctor
Pass the dependencies of VmAnalyzer using its constructor
rather than using setters in order to simplify the code.
Note that some of dependencies can be injected but it since
there might be a lot of VMs running on each host, we prefer
to inject them once in the factory and pass their reference
instead.

Change-Id: Ib347bf5e6da59ce1e3f0fa57130a7e6294c1ff3f
Signed-off-by: Arik Hadas <ahadas@redhat.com>