core: Set command return value when building CoCo command from entity
The CommandCoordinatorImpl.buildCommand() method is called via
retrieveCommand() from several callbacks and the result is used to
persist changes to the command.

In some instances, such as AddDiskCommand, the command may have
previously set a return value that a caller may need to retrieve.  When
the callbacks are executed and persist an update of the returned
command, the return value is cleared.  This causes errors in the calling
code, and in the particular case where this was discovered, makes Image
Upload unable to identify its newly-created disk image.

Change-Id: Idf1067fffe69ed69d45606c0e656f06772864754
Signed-off-by: Greg Padgett <gpadgett@redhat.com>