core: verify Live Merge image deletion on SPM
Bug 1227497 described a case where the engine may forget that it has
removed an image during Live Merge, causing a difficult-to-recover
failure.

The fix submitted did not work for block storage where the VM was
running on a host other than the SPM, because the GetVolumeInfo verb
used is not aware of block storage changes that take place on the SPM.

To avoid this, we created a new SPM GetVolumeInfo command.

Change-Id: Ie14a78a59312c3846be8dc3b1530aa5535c9fb9b
Bug-Url: https://bugzilla.redhat.com/1302215
Related-To: https://bugzilla.redhat.com/1227497
Signed-off-by: Greg Padgett <gpadgett@redhat.com>
Signed-off-by: Ala Hino <ahino@redhat.com>