tools: improved helptag script to auto-generate
Improved the helptag.py script to do a better diff. It can
now read in an existing file containing help URL mappings and
write out a new, updated file while still preserving the URLs.
This saves the team a lot of time.

Removed the JSON format and '--format' option because JSON
support was not actually used.

Change-Id: I07a86ea74c884028cc7a65bd529898d698d64fa0
Signed-off-by: Greg Sheremeta <gshereme@redhat.com>