core: do not validate free mac addresses when importing template
When importing template it's not needed to make sure there are enough
of macs in pool. Until we create VM out of this template no nics will
be needed, so this too-early check may falsely succeed and falsely
fail.

Also when pool is also bound to VmPool check of free macs for template
would not be possible(or at least sane), since we do not know, what
pool will be used, and we may end up checking in wrong pool.

Change-Id: I1c568ce28dc901ef89416473ef0cf960f3276b1e
Signed-off-by: Martin Mucha <mmucha@redhat.com>