core: Live Merge's check for merge completion is wrong
When the DESTROY_IMAGE_CHECK step was added to the Live Merge flow, the
completedMerge() method was not updated.  This method is called upon
failure and checks if the merge operation was completed based on the
current command step.

Without this, a failure during the DESTROY_IMAGE_CHECK step will result
in live merge updating the database in such a way that recovery will
retry the merge operation, requiring manual intervention to repair.

This patch also includes a change to a log message so that it is
possible to tell which database updates the flow performed based on
the result of completedMerge().

Change-Id: Ia20aab2233493b8653c6a714c8504964f8196878
Bug-Url: https://bugzilla.redhat.com/1306367
Signed-off-by: Greg Padgett <gpadgett@redhat.com>