webadmin: Add a scroll panel to SetupNetworks networks list
When a data center contains more then 6 logical networks, the entire
contents of the Setup Host Networks popup dialog would scroll. Local
scrolling on the logical networks list is required to allow both every
logical network and every interface to be in scrolled into view.

This change will allow every logical network to be able to be assigned
to any interface by scrolling both lists as necessary.

Note: VerticalPanels have been replaced with FlowPanels to reduce the
complexity of the generated HTML.

Change-Id: I6466dace1f389ec2a0ecf5789a4ee0ecbd0eeed5
Bug-Url: https://bugzilla.redhat.com/1249895
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>