core: Set unmanaged commands status on engine start
When the engine is shutdown when there are running commands, when the engine
starts again the status of those commands will remain as 'ACTIVE' forever as they
aren't managed by the ATM or by the Coco framework. This may cause to their parent
commands to never end.

This patch adds that responsibllity to the CommandsRepository, as soon as its initiated it'll
update the status of those commands to be ENDED_WITH_FAILURE.

Change-Id: Ib1da52c611fc4171b31d9109b599e4688e7e2b62
Signed-off-by: Liron Aravot <laravot@redhat.com>