core: Enable fencing of previous host which HE VM was running on
If host (for example hostA), which hosted engine VM was running on,
crashes, then hosted engine VM is restarted on another host (hostB).
When engine starts up on hostB, hostA is detected as non responsive, but
it's not fenced automatically because fencing is disabled for 5 minutes
after engine startup (config value DisableFenceAtStartupInSec). This
causes issues:

 1. HA VMs executed on hostA are not restarted on different host
 2. If hostA was SPM, new SPM cannot be selected

This patch enables fencing of hostA even during
DisableFenceAtStartupInSec timeout to prevent above issues.

Change-Id: Iff4368d02d3cbb71dd3cff9c2fc62586fa560a14
Bug-Url: https://bugzilla.redhat.com/1266099
Signed-off-by: Martin Perina <mperina@redhat.com>