restapi: Make supported and default versions configurable
This patch adds two new parameters to the engine local configuration
file:

ENGINE_API_DEFAULT_VERSION

  Comma separated list of supported versions of the API. When the user
  selects a version of the API that isn't in this list the server will
  return an HTTP error response.

  The default value will be "4".

ENGINE_API_SUPPORTED_VERSIONS

  Default version of the API, which will be used when the caller
  doesn't explicitly select a specific version.

  The default value will be "3, 4".

These parameters, specially the one that sets the default version, will
be useful for environments where keeping backwards compatibility is
paramount. For example, if there are many API clients in the environment
that use version 3 of the API it is simpler to change the default used
by the server than to change all the clients to explicitly request
version 3 (using the "Version" header or the "/v3" URL prefix).

Change-Id: I876e5c1c1c9b626e5ce1e173a3c0b5555561121b
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>