webadmin: Network Import dialog open/cancel fails
The Network Import dialog in the network list fails to
close when cancel is pressed. It also opens only once,
all further attempts to open it fail.
The issue was caused by an incorrect model being injected
into the model (the external provider one).
The same dialog is also used in external providers network
import. There it works fine, as the model is the right one.

Change-Id: Id6a687c0c290891b3c5636dc82d598399f3e37df
Bug-Url: https://bugzilla.redhat.com/1306585
Signed-off-by: mirecki <mmirecki@redhat.com>